% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSPIRE.R
\docType{data}
\name{exmp_dataset1}
\alias{exmp_dataset1}
\title{Example Gene Expression Dataset-1}
\description{
This example ovarian cancer dataset contains expression of random half of the genes
on the 28 samples from the GSE19829.GPL570 accession in Gene Expression Omnibus.
Contains 28 samples (as rows) and 9056 genes (as columns).
4117 of the genes are overlapping with the genes in exmp_dataset2.
}

