% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSPIRE.R
\docType{data}
\name{exmp_dataset2}
\alias{exmp_dataset2}
\title{Example Gene Expression Dataset-2}
\description{
This example ovarian cancer dataset contains expression of random half of the genes
on the 42 samples from the GSE19829.GPL8300 accession in Gene Expression Omnibus.
Contains 42 samples (as rows) and 4165 genes (as columns).
4117 of the genes are overlapping with the genes in exmp_dataset1.
}

