% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspire_me.R
\name{inspire_me}
\alias{inspire_me}
\title{Get some data storytelling inspiration}
\usage{
inspire_me(author = "", affiliation = "", topic = "", type = "")
}
\arguments{
\item{author}{optional parameter to specify desired author of returned content}

\item{affiliation}{optional parameter to specify desired affiliation of content}

\item{topic}{optional parameter to specify desired topic of returned content}

\item{type}{optional parameter to specify desired type of returned content}
}
\value{
A URL
}
\description{
This functions returns a random URL of data storytelling inspiration based on our contributed list. Users can specify author, affiliation, topic, or type if they wish.
}
